DynaWaltz: network blackbox models do not have staticRef
