+++ dynawaltz/src/main/java/com/powsybl/dynawaltz/dynamicmodels/AbstractBlackBoxModelWithStaticId.java
@@ -1,0 +1,16 @@
+package com.powsybl.dynawaltz.dynamicmodels;
+
+public abstract class AbstractBlackBoxModelWithStaticId extends AbstractBlackBoxModel implements BlackBoxModelWithStaticId {
+
+    private final String staticId;
+
+    protected AbstractBlackBoxModelWithStaticId(String dynamicModelId, String staticId, String parametersId) {
+        super(dynamicModelId, parametersId);
+        this.staticId = staticId;
+    }
+
+    @Override
+    public String getStaticId() {
+        return staticId;
+    }
+}
