+++ dynawaltz/src/main/java/com/powsybl/dynawaltz/dynamicmodels/BlackBoxModelWithStaticId.java
@@ -1,0 +1,5 @@
+package com.powsybl.dynawaltz.dynamicmodels;
+
+public interface BlackBoxModelWithStaticId extends BlackBoxModel{
+    String getStaticId();
+}
