+++ dynawaltz/src/main/java/com/powsybl/dynawaltz/dynamicmodels/nonstaticref/network/AbstractNetworkBlackBoxModel.java
@@ -1,9 +1,9 @@
 package com.powsybl.dynawaltz.dynamicmodels.nonstaticref.network;
 
 import com.powsybl.dynawaltz.DynaWaltzContext;
+import com.powsybl.dynawaltz.dynamicmodels.AbstractBlackBoxModelWithStaticId;
 import com.powsybl.dynawaltz.dynamicmodels.BlackBoxModel;
 import com.powsybl.dynawaltz.dynamicmodels.MacroConnector;
-import com.powsybl.dynawaltz.dynamicmodels.staticref.BlackBoxModelWithStaticRef;
 import com.powsybl.dynawaltz.xml.DynaWaltzXmlConstants;
 import org.apache.commons.lang3.tuple.Pair;
 
@@ -12,27 +12,10 @@ import javax.xml.stream.XMLStreamWriter;
 import java.util.Collections;
 import java.util.List;
 
-public abstract class AbstractNetworkBlackBoxModel implements BlackBoxModelWithStaticRef {
-
-    private final String staticId;
+public abstract class AbstractNetworkBlackBoxModel extends AbstractBlackBoxModelWithStaticId {
 
     protected AbstractNetworkBlackBoxModel(String staticId) {
-        this.staticId = staticId;
-    }
-
-    @Override
-    public String getStaticId() {
-        return staticId;
-    }
-
-    @Override
-    public String getDynamicModelId() {
-        return "";
-    }
-
-    @Override
-    public String getParameterSetId() {
-        return "";
+        super("", staticId, "");
     }
 
     @Override
@@ -48,11 +31,6 @@ public abstract class AbstractNetworkBlackBoxModel implements BlackBoxModelWithS
     @Override
     public void writeParameters(XMLStreamWriter writer, DynaWaltzContext context) throws XMLStreamException {
         // Default models not written in dyd
-    }
-
-    @Override
-    public List<Pair<String, String>> getStaticRef() {
-        return Collections.emptyList();
     }
 
     @Override
