+++ dynawaltz/src/main/java/com/powsybl/dynawaltz/dynamicmodels/staticref/AbstractBlackBoxModelWithStaticRef.java
@@ -7,7 +7,7 @@
 package com.powsybl.dynawaltz.dynamicmodels.staticref;
 
 import com.powsybl.dynawaltz.DynaWaltzContext;
-import com.powsybl.dynawaltz.dynamicmodels.AbstractBlackBoxModel;
+import com.powsybl.dynawaltz.dynamicmodels.AbstractBlackBoxModelWithStaticId;
 import com.powsybl.dynawaltz.xml.MacroStaticReference;
 
 import javax.xml.stream.XMLStreamException;
@@ -18,17 +18,10 @@ import static com.powsybl.dynawaltz.xml.DynaWaltzXmlConstants.DYN_URI;
 /**
  * @author Luma Zamarreño <zamarrenolm at aia.es>
  */
-public abstract class AbstractBlackBoxModelWithStaticRef extends AbstractBlackBoxModel implements BlackBoxModelWithStaticRef {
-
-    private final String staticId;
+public abstract class AbstractBlackBoxModelWithStaticRef extends AbstractBlackBoxModelWithStaticId implements BlackBoxModelWithStaticRef {
 
     protected AbstractBlackBoxModelWithStaticRef(String dynamicModelId, String staticId, String parameterSetId) {
-        super(dynamicModelId, parameterSetId);
-        this.staticId = staticId;
-    }
-
-    public String getStaticId() {
-        return staticId;
+        super(dynamicModelId, staticId, parameterSetId);
     }
 
     protected void writeBlackBoxModel(XMLStreamWriter writer, DynaWaltzContext context) throws XMLStreamException {
